Matters to categorize the pictures

+++ db/migrate/20121013000512_create_helpers.rb
@@ -2,7 +2,7 @@ class CreateHelpers < ActiveRecord::Migration
   def change
     create_table :helpers do |t|
       t.string :name
-      t.float :price
+      t.integer :price
       t.string :identifier
 
       t.timestamps
